Add a list of taxonomic updates into a metadata file for a dataset_id

Add multiple taxonomic changes to the metadata yaml file using a dataframe containing the taxonomic changes to be made.

metadata_add_taxonomic_changes_list(dataset_id, taxonomic_updates)

Arguments

dataset_id

identifier for a particular study in the AusTraits database

taxonomic_updates

dataframe of taxonomic updates

Value

yml file with multiple taxonmic updates added